김성훈 : 엔진 코드 업데이트(엑터 오더 관리, 시간과 상호작용하는 애니메이션 코드 추가, PrevLevel정보 받아오기)

+++ API/GameEngineContents/Library.cpp
@@ -3,7 +3,7 @@
 #include <GameEngine/GameEngineRenderer.h>
 #include <GameEngine/GameEngineCollision.h>
 
-int TelePortX = 3584;
+float TelePortX = 3584.0f;
 
 Library::Library() 
 {
